deletar perfis

+++ app/Http/Controllers/Admin/PlanController.php
@@ -45,10 +45,19 @@ class PlanController extends Controller
     }
 
     public function destroy($id){
-        $plan = $this->repository->where('url', $id)->first();
+        $plan = $this->repository
+                            ->with('details')
+                            ->where('url', $id)
+                            ->first();
 
         if(!$plan){
             return redirect()->back();
+        }
+
+        if($plan->details->count() > 0){
+            return redirect()
+                        ->back()
+                        ->with('error', 'Existem detalhes vinculados a esse plano, portanto não pode deletar');
         }
 
         $plan->delete();
